Split NetcdfFileReader into separate file. Add some doc strings.
